# src/BackendUsage/Objects/PlayerProfileDTO.cs
using System;

namespace CompetitiveBackend.BackendUsage.Objects
{
    [Serializable]
    public class PlayerProfileDTO : IntIdentifiableDTO
    {
        public string? Name { get; set; }
        public string? Description { get; set; }
        public PlayerProfileDTO(string? name, string? description, int? iD)
            : base(iD)
        {
            Name = name;
            Description = description;
        }
    }
}