# www/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CompetitiveBackend",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/competitions/{compID}/game_session": {
      "get": {
        "tags": [
          "CompetitionRiddleGameplay"
        ],
        "summary": "Получить задание на игру",
        "parameters": [
          {
            "name": "compID",
            "in": "path",
            "description": "Идентификатор соревнования",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное выполнение",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompetitionParticipationTaskDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetitionParticipationTaskDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetitionParticipationTaskDTO"
                }
              }
            }
          },
          "404": {
            "description": "Соревнование с таким ID не найдено",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Не является игроком",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Ошибка сервера"
          }
        }
      },
      "post": {
        "tags": [
          "CompetitionRiddleGameplay"
        ],
        "summary": "Выполнить задание на игру",
        "parameters": [
          {
            "name": "compID",
            "in": "path",
            "description": "Идентификатор соревнования",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Ответы на вопросы",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/CompetitionParticipationRequestDTO"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompetitionParticipationRequestDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CompetitionParticipationRequestDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CompetitionParticipationRequestDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешное выполнение",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipationFeedbackDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipationFeedbackDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipationFeedbackDTO"
                }
              }
            }
          },
          "404": {
            "description": "Соревнование с таким ID не найдено",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Не игрок",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Ошибка сервера"
          }
        }
      }
    },
...
    "/api/v1/players/{profileID}": {
      "get": {
        "tags": [
          "Player"
        ],
        "summary": "Получить информацию о профиле",
        "parameters": [
          {
            "name": "profileID",
            "in": "path",
            "description": "Идентификатор игрока",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerProfile"
                }
              }
            }
          },
          "500": {
            "description": "Ошибка сервера"
          }
        }
      }
    },
...
    "/api/v1/competitions/{compID}/participations/{profileID}": {
      "get": {
        "tags": [
          "PlayerParticipation"
        ],
        "summary": "Получить полную информации об участии в соревновании",
        "parameters": [
          {
            "name": "profileID",
            "in": "path",
            "description": "Идентификатор профиля",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "compID",
            "in": "path",
            "description": "Идентификатор соревнования",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное выполнение",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerParticipation"
                }
              }
            }
          },
          "404": {
            "description": "Участие пользователя profileID в соревновании compID не найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Ошибка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "PlayerParticipation"
        ],
        "summary": "Удалить участие игрока в соревновании",
        "parameters": [
          {
            "name": "profileID",
            "in": "path",
            "description": "Идентификатор профиля игрока",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "compID",
            "in": "path",
            "description": "Идентификатор соревнования",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Участие успешно удалено"
          },
          "404": {
            "description": "Участие не найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Ошибка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompetitionParticipationRequestDTO": {
        "type": "object",
        "properties": {
          "sessionID": {
            "type": "string",
            "nullable": true
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RiddleAnswerDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CompetitionParticipationTaskDTO": {
        "type": "object",
        "properties": {
          "sessionID": {
            "type": "string",
            "nullable": true
          },
          "riddles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserRiddleInfoDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ParticipationFeedbackDTO": {
        "type": "object",
        "properties": {
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "rightAnswersCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalAnswersCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "PlayerParticipation": {
        "type": "object",
        "additionalProperties": false
      },
...
      "PlayerProfile": {
        "type": "object",
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "RiddleAnswerDTO": {
        "type": "object",
        "properties": {
          "textAnswer": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UserRiddleInfoDTO": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string",
            "nullable": true
          },
          "availableAnswers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RiddleAnswerDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Put **_ONLY_** your JWT Bearer token on textbox below!",
        "name": "Bearer",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}